% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{as_list}
\alias{as_list}
\title{Convert an VOR object to a list}
\usage{
VORobj$as_list()
}
\value{
A list
}
\description{
This method extracts all fields of an VOR object and places their stored values 
into the fields of an R list, with list field names matching the VOR object field names.
}
