% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{calc_all}
\alias{calc_all}
\title{Calc all Voronoi quantities}
\usage{
VORobj$calc_all()
}
\description{
This wrapper method computes all Voronoi quantities exposed by the VOR class, in the following order: 
\itemize{
\item{\code{calc_DADJ}}
\item{\code{calc_vor2_centers}}
\item{\code{calc_vor1_Dikin}}
\item{\code{calc_vor2_Dikin}}
\item{\code{calc_vor1_MVIE}}
\item{\code{calc_vor2_MVIE}}
}
}
