% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{calc_DADJ}
\alias{calc_DADJ}
\title{Calculate the Delaunay graph adjacency}
\usage{
VORobj$calc_DADJ
}
\value{
None, the field \code{DADJ} is set internally.
}
\description{
The Delaunay graph dual of a first-order Voronoi tessellation has edges between 
prototype vertices whose corresponding Voronoi cells intersect (share a face). This method 
computes the (binary) adjacency matrix of this graph using the method Agrell (referenced below),  
which involves solving a linear program for each unique pair of prototypes in the tessellation. 
To speed up the computation, the CADJ and Gabriel adjacency matrices (set during \code{initialize_VOR}) are 
used to "seed" the Delaunay adjacency (as proper sub-graphs of the Delaunay graph, the existence of an edge between 
prototypes \code{i} and \code{j} in either necessarily indicates a corresponding edge in DADJ).  
Note: a Delaunay adjacency is required for all ellipsoidal calculations in VorVQ. 
As the many thousand LPs required for full DADJ calculation can be quite large it is recommended 
to perform this calculation in parallel, which can be set in \link[VorVQ]{set_params} (default is \code{parallel = TRUE}).
}
\references{
\insertRef{Agrell1993}{VorVQ}
}
