% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{calc_GADJ}
\alias{calc_GADJ}
\title{Calculate the Gabriel graph adjacency}
\usage{
VORobj$calc_GADJ()
}
\value{
None
}
\description{
The Gabriel graph is a proximity graph that is a known sub-graph of the Delaunay triangulation 
induced by the Voronoi tessellation of the prototypes in \code{W}.  The Gabriel ADJacency matrix helps 
speed up the computation of the full Delaunay graph and is calculated automatically during \code{initialize_VOR} 
via an internal call to \code{calc_GADJ}. This method should not need to be called directly by a user but it 
exposed for completeness.
}
\details{
Field \code{GADJ} will be overwritten internally.
}
\references{
\insertRef{Gabriel1969}{VorVQ}
}
