% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{calc_vor1_centers}
\alias{calc_vor1_centers}
\title{Calculate the Chebyshev centers of first-order Voronoi cells}
\usage{
VORobj$calc_vor1_centers
}
\value{
None, the field \code{vor1_centers} is set internally.
}
\description{
The \href{https://en.wikipedia.org/wiki/Chebyshev_center}{Chebyshev center} of the polytope definition 
(from \code{get_vor1_polytope}) for each **active** first-order Voronoi cell (as specified in \code{vor1_active}) is computed. 
If set, these centers are used as starting interior points for MVIE and Dikin ellipsoid calculations. 
Clear any previously computed centers via \code{clear_vor1_centers}.
}
