% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{calc_vor1_MVIE}
\alias{calc_vor1_MVIE}
\title{Compute the MVIE for first-order Voronoi cells}
\usage{
VORobj$calc_vor1_MVIE()
}
\value{
None, the following fields store components of the MVIE: 
\describe{
\item{\code{vor1_MVIE_c}}{matrix of ellipsoid centers, rows correspond to entries in \code{vor1_active}}
\item{\code{vor1_MVIE_E}}{cube of ellipsoid rotation matrices, slices correspond to entries in \code{vor1_active}}
\item{\code{vor1_MVIE_logdet}}{vector of the log-determinant of the ellipsoid rotations stored in the slices of \code{vor1_MVIE_E}}
\item{\code{vor1_MVIE_status}}{vector of the return flags from calling the MVIE routinefor each \code{vor1_active} cell}
\item{\code{vor1_MVIE_volratio}}{vector of the proportional volume of each \code{vor1_active} MVIE (should sum to unity)}
}
}
\description{
The **M**aximum **V**olume **I**nscribed **E**llipsoid inside the polytope defined by each 
active first-order Voronoi cell is computed and stored, according to the method of Zhang & Gao. 
The starting points for the routine are taken from \code{vor1_centers}, if it has been set 
(via either \code{calc_vor1_centers} or \code{set_vor1_centers}); otherwise the prototypes stored in \code{W} 
are used.  The parameters controlling the routine can be viewed or changed via \code{get/set_params}. 
It is recommended to allow parallel computation for MVIE calculation (i.e., \code{set_params(parallel = TRUE)}.
}
\details{
Details of the MVIE routine can be found in the help of \link[VorVQ]{max_vol_inscr_ell}.
}
\references{
\insertRef{ZhangGao2003}{VorVQ}
}
