% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{calc_vor2_centers}
\alias{calc_vor2_centers}
\title{Calculate the Chebyshev centers of second-order Voronoi cells}
\usage{
VORobj$calc_vor2_centers
}
\value{
None, the field \code{vor2_centers} is set internally.
}
\description{
The \href{https://en.wikipedia.org/wiki/Chebyshev_center}{Chebyshev center} of the polytope definition 
(from \code{get_vor2_polytope}) for each **active** second-order Voronoi cell (as specified in \code{vor2_active}) is computed. 
These centers are used as starting interior points for MVIE and Dikin ellipsoid calculations. 
Clear any previously computed centers via \code{clear_vor2_centers}.
}
