% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{calc_vor2_Dikin}
\alias{calc_vor2_Dikin}
\title{Compute the Dikin ellipsoid for second-order Voronoi cells}
\usage{
VORobj$calc_vor2_Dikin()
}
\value{
None, the field \code{vor2_MVIE_E}, which is a cube whose slices contain the \code{vor2_active} Dikin ellipsoid 
rotations, is stored internally.
}
\description{
The Dikin ellipsoid is a locally inscribed geometric approximator of a polytope.  
These ellipsoidal approximators are computed for each active second-order Voronoi cell 
at the points specified in \code{vor2_centers}
}
\details{
Details of the Dikin ellipsoid can be found in the help of \link[VorVQ]{Dikin_Ellipsoid}.
}
\references{
\insertRef{Dikin1967}{VorVQ}
\insertRef{Boyd2004}{VorVQ}
}
