% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{clear_vor2_centers}
\alias{clear_vor2_centers}
\title{Clear the centers of second-order Voronoi cells}
\usage{
VORobj$clear_vor2_centers
}
\value{
None, the field \code{vor2_centers} is reset to a matrix with \code{nrows = ncols = 0}.
}
\description{
This method clears any values stored in the field \code{vor2_centers} that were previously populated by 
either \code{calc_vor2_centers} or \code{set_vor2_centers}.
}
