% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{CMIX9}
\alias{CMIX9}
\title{A Synthethic 2-D Gaussian Mixture}
\format{
These data are stored in a list variable named \code{CMIX9} with components: 
\describe{
\item{X}{data matrix with 5000 rows and 2 columns}
\item{W}{matrix of 100 learned prototype vectors (in rows, learned from a 10x10 SOM)}
\item{CADJ}{The CADJ matrix resulting from a recall of the data in \code{X} using the prototypes in \code{W}. 
\code{CADJ} is ordered such that its \code{(i,j)} element contains the number of data vectors in \code{X} whose 
BMU1 = the prototype vector in the \code{i}-th row and BMU2 = the prototype in the \code{j}-th row of \code{W}}
}
}
\usage{
CMIX9
}
\description{
Table 1 of \insertCite{Chacon2009}{VorVQ} compiles twelve different two-dimensional synthetic Gaussian 
mixtures which have been exercised extensively for the task of density estimation. For demonstration of 
Voronoi calculations and visualization in lower dimension we provide Mixture 9 in \code{VorVQ}.
}
\references{
\insertRef{Chacon2009}{VorVQ}
}
\keyword{datasets}
