% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Delaunay_ADJ}
\alias{Delaunay_ADJ}
\title{Build the Delaunay Adjacency Matrix of a set of prototypes}
\usage{
Delaunay_ADJ(W, lb, ub, parallel = TRUE)
}
\arguments{
\item{W}{prototype matrix, with prototype vectors in rows}

\item{lb}{a vector of the global (data range) lower bounds, by dimension}

\item{ub}{a vector of the global (data range) upper bounds, by dimension}

\item{parallel}{whether to process in parallel. Default = TRUE.}
}
\value{
A (binary, symmetric) adjacency matrix, nrows = ncols = nrow(W), giving
the Delaunay adjacencies between all prototypes in W. \cr
The global lower and upper bounds (xlb and xub) are needed to ensure the polyope definition is bounded.
Otherwise, the LP we need to solve here could be unbounded, which produces incorrect results.
}
\description{
Build the Delaunay Adjacency Matrix of a set of prototypes
}
\references{
\insertRef{Delaunay1934}{VorVQ}

\insertRef{Agrell1993}{VorVQ}
}
