% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Dikin_ell}
\alias{Dikin_ell}
\title{Find Dikin ellipsoids inside a polytope}
\usage{
Dikin_ell(A, b, c)
}
\arguments{
\item{A}{LHS of polytope constraints}

\item{b}{RHS of polytope constraints}

\item{c}{interior point where Dikin ellipse is centered}
}
\value{
the rotation matrix of the Dikin ellipsoid centered at \code{c}.
}
\description{
Returns a matrix E such that the ellipsoid \eqn{(y - c)'E^-1 (y - c) <= 1} is inscribed in the polytope defined by Ax <= b, 
with curvature influenced by the local geometry of the constraints at \eqn{c}
}
\references{
\insertRef{Dikin1967}{VorVQ}
\insertRef{Boyd2004}{VorVQ}
}
