% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gabriel_ADJ}
\alias{Gabriel_ADJ}
\title{Compute the Gabriel graph adjacency}
\usage{
Gabriel_ADJ(W, parallel = TRUE)
}
\arguments{
\item{W}{the prototype matrix, prototype vectors in rows}

\item{parallel}{boolean, whether to compute in parallel. Default = TRUE.}
}
\value{
a (binary, symmetric) adjacency matrix of dimension \code{nrow(W) x nrow(W)}.  Element 
\code{(i,j) = 1} IFF the prototype vectors \code{W[i,]} and \code{W[j,]} are Gabriel adjacent.  
Otherwise, entries are 0.
}
\description{
Compute the Gabriel graph adjacency
}
\details{
The Gabriel graph is a proximity graph of points that is known to be a sub-graph 
of the Delaunay triangulation of the same set of points.
}
\references{
\insertRef{Gabriel1969}{VorVQ}
}
