% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{get_params}
\alias{get_params}
\title{Get parameters for Voronoi calculations}
\usage{
VORobj$get_params()
}
\value{
A list with named components equal to the parameters.
}
\description{
Several methods require parameters whose values can be accessed via this method. 
See \link[VorVQ]{set_params} for a description of each.
}
