% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{get_vor1_centers}
\alias{get_vor1_centers}
\title{Get the centers of first-order Voronoi cells}
\usage{
VORobj$get_vor1_centers
}
\description{
If \code{calc_vor1_centers} or \code{set_vor1_centers} has been set, this method returns the centers 
calculated or set by these methods. Otherwise, the \code{vor1_active} prototypes in \code{W} are returned.
}
