% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{get_vor1_polytope}
\alias{get_vor1_polytope}
\title{Get the definition of a first-order Voronoi cell}
\usage{
VORobj$get_vor1_polytope(iidx)
}
\arguments{
\item{iidx}{an integer identifying which first-order Voronoi cell's definition is returned. 
It should correspond to the row index of the generator in \code{W}.}
}
\value{
A list with components: 
\describe{
\item{A}{The LHS coefficient matrix of the linear inequality system}
\item{b}{The RHS upper bounds of the linear system, length = \code{nrow(A)}}
\item{cid}{A vector of indices (length = \code{nrow(A)}) indicating which prototype (generator) 
induced the half-plane constraint stored in the corresponding row of \code{A} and \code{b}. 
Constraints involving the global bounds have \code{cid = nrow(W)+1} by convention.}  
}
}
\description{
The definition of a first-order Voronoi cell naturally induces a half-plane representation 
of each cell, of the form \eqn{Ax <= b}. This method will return this half-plane definition 
for a particular cell. 
By convention, the dimension-wise lower and upper bounds stored in \code{lb} and \code{ub} 
are appended to the system to ensure the resulting polytope is closed. 
Since constraints arising from any non-Delaunay adjacent prototypes (as defined in \code{DADJ}) are redundant in the above 
system, these are not included in the returned polytope definition (making it non-redundant, as small as possible).
}
\references{
\insertRef{Agrell1993}{VorVQ}
}
