% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{get_vor2_centers}
\alias{get_vor2_centers}
\title{Get the centers of second-order Voronoi cells}
\usage{
VORobj$get_vor2_centers
}
\description{
Returns the second-order centers that were set when calling either \code{calc_vor2_centers} or \code{set_vor2_centers}.
}
