% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{get_vor2_polytope}
\alias{get_vor2_polytope}
\title{Get the definition of a second-order Voronoi cell}
\usage{
VORobj$get_vor2_polytope(iidx, jidx)
}
\arguments{
\item{iidx}{}

\item{jidx}{}
}
\value{
A list with components: 
\describe{
\item{A}{The LHS coefficient matrix of the linear inequality system} 
\item{b}{The RHS upper bounds of the linear system, length = \code{nrow(A)}} 
\item{cid}{A vector of indices (length = \code{nrow(A)}) indicating which prototype (generator) 
induced the half-plane constraint stored in the corresponding row of \code{A} and \code{b}. 
Constraints involving the global bounds have \code{cid = nrow(W)+1} by convention.} 
}
}
\description{
The definition of a second-order Voronoi cell naturally induces a half-plane representation 
of each cell, of the form \eqn{Ax <= b}. This method will return this half-plane definition 
for a particular cell. 
By convention, the dimension-wise lower and upper bounds stored in \code{lb} and \code{ub} 
are appended to the system to ensure the resulting polytope is closed. 
Since constraints arising from any non-Delaunay adjacent prototypes (as defined in \code{DADJ}) are redundant in the above 
system, these are not included in the returned polytope definition (making it non-redundant, as small as possible).
}
\details{
\code{iidx} and \code{jidx} are integer indices (to the generators specified in the rows of \code{W}) 
which, together, define the second-order Voronoi cell \code{iidx-jidx}.
}
\references{
\insertRef{Agrell1993}{VorVQ}
}
