% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_Delaunay_nhb}
\alias{is_Delaunay_nhb}
\title{Determine if two prototypes are Delaunay neighbors}
\usage{
is_Delaunay_nhb(W, iidx, jidx, lb, ub, verbose = TRUE)
}
\arguments{
\item{W}{prototype matrix, with prototype vectors in rows}

\item{iidx}{the first index of the neighbor pair to test}

\item{jidx}{the second index of the neighbor pair to test}

\item{lb}{a vector of the global (data range) lower bounds, by dimension}

\item{ub}{a vector of the global (data range) upper bounds, by dimension}

\item{verbose}{boolean, whether to display the LP solver status}
}
\value{
1 = prototypes iidx and jidx are neighbors, 0 = they are not
}
\description{
Determine if two prototypes are Delaunay neighbors
}
\details{
iidj and jidx should be valid row indices of W indexed from 1 (not 0). \cr
The global lower and upper bounds (xlb and xub) are needed to ensure the polyope definition is bounded.
Otherwise, the LP we need to solve here could be unbounded, which produces incorrect results.
}
\references{
\insertRef{Delaunay1934}{VorVQ}

\insertRef{Agrell1993}{VorVQ}
}
