% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_interior_point}
\alias{is_interior_point}
\title{Check if point is interior}
\usage{
is_interior_point(A, b, x)
}
\arguments{
\item{A}{the LHS matrix of polytope constraints}

\item{b}{the RHS vector of polytope constraints}

\item{x}{the query point}
}
\value{
boolean = T if the query point satisfies the above condition
}
\description{
For a feasible set of linear equations Ax<=b that define a bounded polytope P 
and a point x0, x0 is interior IFF b - Ax > 0.
}
