% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{load}
\alias{load}
\title{Load an existing VOR object}
\usage{
VORobj$load(vorfile)
}
\arguments{
\item{vorfile}{a string indicating the file path and name of the saved VOR object.}
}
\value{
None, the VOR object is loaded
}
\description{
VOR objects previously written to disk via the \link{save} method can be re-loaded into 
a new R environment with this function. All fields of the internal C++ class will be populated, and 
all methods can be called on the loaded VOR object.
}
\details{
Because the .vor file is in .rds format it can, technically, be loaded directly into an R environment 
as a list via \link[base]{readRDS}. This can be useful for spot checking the contents of a saved VOR object, but 
does not allow use of any of its methods (or visualizations).  The \code{load} methods allows for 
proper restoration of a previously saved VOR
}
