% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{load_list}
\alias{load_list}
\title{Populate a VOR object from a list}
\usage{
VORobj$load_list(VORList)
}
\arguments{
\item{VORList}{a VOR object converted to a list, e.g., with \code{as_list}}
}
\value{
None
}
\description{
This method populates all fields of a VOR object from the fields of an R list object. 
The list must have field names which exactly match the VOR field names.
}
