% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_vol_inscr_ell}
\alias{max_vol_inscr_ell}
\title{Max volume inscribed ellipsoid of polytope}
\usage{
max_vol_inscr_ell(
  A,
  b,
  c0,
  maxiter = 100L,
  tol = 1e-04,
  fix_c0 = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{A}{matrix of LHS constraints in H-representation of polytope}

\item{b}{vector of RHS constraints in H-representation of polytope}

\item{c0}{a point inside the polytope (not on the boundary)}

\item{maxiter}{integer number of iterations to perform, default = 100}

\item{tol}{tolerance for Zhang's MVE code, see paper, default = 1e-4}

\item{fix_c0}{boolean, whether to fix the given center at c0 during iteration, default = F}

\item{verbose}{boolean, whether to display the solver status at each iteration, default = F}
}
\value{
a list with components
\itemize{
\item{E}{a matrix defining the rotation and scale of the maximum volume ellipsoid}
\item{c}{a vector defining the center of the maximum volume ellipsoid}
\item{status}{a flag returned from Zhang's solver code.  See details.}
}
}
\description{
Returns the matrix E and vector c such that the ellipsoid \eqn{(y - c)'E^-1 (y - c) <= 1} is inscribed in the polytope defined by Ax <= b with maximum volume.
Code taken from Yin Zhang's \verb{MVE} routine.
}
\details{
\verb{maxiter} and \verb{tol} are control (convergence) parameters
for Yin Zhang's \verb{MVE} routine.

The polytope defined by A & b must be bounded.

status = 0 indicates successful convergence; = -1 means maxiter was reached before convergence.
}
\references{
\insertRef{ZhangGao2003}{VorVQ}
}
