% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{new}
\alias{new}
\title{Create an empty VOR object}
\usage{
VOR$new()
}
\value{
an empty VOR object templated for initialization (via \link{initialize_VOR}).
}
\description{
Create an empty VOR object
}
