% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{polytope_Chebyshev_center}
\alias{polytope_Chebyshev_center}
\title{Chebyshev center of a polytope}
\usage{
polytope_Chebyshev_center(A, b)
}
\arguments{
\item{A}{the LHS matrix of polytope constraints}

\item{b}{the RHS vector of polytope constraints}
}
\value{
a vector (of the same dimension as A) giving the interior point
}
\description{
For a feasible set of linear equations Ax<=b that define a bounded polytope P, 
the Chebyshev center is the center the largest inscribed ball contained in P.
Polytope definition should be given as Ax <= b.
}
