% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_polytope.R
\name{polytope_chull}
\alias{polytope_chull}
\title{Compute the convex hull of vertices of a polytope}
\usage{
polytope_chull(A, b)
}
\arguments{
\item{A}{the LHS constraint matrix}

\item{b}{the RHS constraint vector}
}
\value{
a list with components: 
\enumerate{
\item V a matrix with vertex points in its rows
\item Vpath a 2 column edge list matrix whose rows specify vertex indices (index to rows of V)
  that are connected in the convex hull path. Note that this path has its redundancies removed 
  (i.e., a connection between vertex 1 & 2 is given only by a row (1,2), not by (2,1))
}
}
\description{
Polytope definition should be given in the H-representation Ax <= b. 
This function uses Komei Fukuda's cdd library to convert it to a V-representation.
}
