% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{save}
\alias{save}
\title{Save a VOR object}
\usage{
VORobj$save(vorfile)
}
\arguments{
\item{vorfile}{a string indicating the file path and name in which to save the VOR object.
This must end in extension ".vor", otherwise an error is returned.}
}
\value{
None, the VOR object is saved to disk
}
\description{
All fields in a VOR object can be saved to disk with this function, which allows them to be re-loaded 
into a new R environment at a later time (for analysis, or possibly extended training).
}
\details{
The VOR object is saved to disk as an R list, with each field occupying a corresponding field of the list.  
The file is saved in .rds format (can check its details with \link[base]{infoRDS}).  

Saved VORs can be re-loaded with \link{load}
}
