% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{set_bounds}
\alias{set_bounds}
\title{Set the bounds of the tessellated region}
\usage{
VORobj$set_bounds(lb, ub)
}
\arguments{
\item{lb}{a vector of lower bounds}

\item{ub}{a vector of upper bounds}
}
\value{
None
}
\description{
To ensure that all Voronoi cells are closed in \eqn{R^d} dimension-wise lower and upper 
bounds of the entire tessellated region must be specified. By default, the upper and lower bounds 
for each dimension are set = +/- 5% of \code{range(W)} in \code{initialize_VOR}.  This method is used 
to change these defaults to a user specified range, which must span the entire range of W 
in each dimension. This will be checked internally. Note that changing the global bounds can 
alter the Delaunay edges between prototypes which are on/near the perimiter of the point cloud defined by \code{W}.
}
\details{
Both \code{lb} and \code{ub} can have \code{length = 1}, in which case the single 
value will be recycled across dimension. Otherwise, they must have \code{length = ncol(W)}. 
Both must be supplied, even if only 
Internally, fields \code{lb} and \code{ub} will be overwritten.
}
