% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{set_DADJ}
\alias{set_DADJ}
\title{Set the Delaunay graph adjacency}
\usage{
VORobj$set_DADJ
}
\value{
None, the field \code{DADJ} is set internally.
}
\description{
A Delaunay graph adjacency is required for all ellipsoidal calculations in VorVQ. 
If the Delaunay adjacency is available from an external source (or calculation), it can 
be set with this method. Otherwise, one can be computed (and set) via \link[VorVQ]{calc_DADJ}.
}
