% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{set_params}
\alias{set_params}
\title{Set parameters for Voronoi calculations}
\usage{
VORobj$set_params(list(param_name = param_value))
}
\arguments{
\item{param_list}{a list with component names in the parameter set defined above, and corresponding desired values.}
}
\value{
None
}
\description{
Several methods require parameters which can be accessed via this method. 
The parameters and their defaults (set during a call to \code{initialize_VOR} are: 
\describe{
\item{\code{parallel}}{boolean, whether to perform all calculations in parallel (via RcppParallel). 
Default = T.}
\item{\code{MVIE_maxiter}}{The maximum number of iterations allowed when calculating the MVIEs for Voronoi cells. Default = 100.}
\item{\code{MVIE_tol}}{The relative convergence tolerance for the MVIE routine, which exits when the volume of the computed MVIE changes < this tol. Default = 1e-4.}
\item{\code{MVIE_fix_c0}}{boolean, whether to allow the MVIE routine to update the MVIE center along with the ellipsoidal rotation. Default = T.}
}
}
