% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{set_vor1_active}
\alias{set_vor1_active}
\title{Set the active first-order Voronoi cells}
\usage{
VORobj$set_vor1_active(vor1_indices)
}
\arguments{
\item{vor1_indices}{a vector containing the indices (rows of \code{W}, 1-based) of the desired active first-order cells}
}
\value{
None, the field \code{vor1_active} is set internally.
}
\description{
The computations with a VOR object (computing Voronoi centers and ellipsoids) are only performed for "active" cells, 
which are identified by the prototype index (row index in \code{W}) which generated the cell. 
By default the active first-order Voronoi cells are those which contain data mapped to them during a recall 
of the ANN, as indicated by CADJ (the first-order cells corresponding to the rows of CADJ with with non-zero rowsum). 
A different set of active cells can be set by calling this method
}
