% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{set_vor1_centers}
\alias{set_vor1_centers}
\title{Set the centers of first-order Voronoi cells}
\usage{
VORobj$set_vor1_centers
}
\arguments{
\item{C}{a matrix whose rows specify the active first-order Voronoi cell centers. Must have \code{nrow(C) = length(vor1_active)}.}
}
\value{
None, the field \code{vor1_centers} is set internally.
}
\description{
The MVIE and Dikin methods require a starting interior point for each active first-order Voronoi polytope. 
Be default, the prototypes (Voronoi cell generators, as stored in \code{W}) of each active first-order cell is used. 
Different interior points can be set with this method. 
Note: the points passed to this method **must be interior** to each first-order cell; a check will 
be performed and an error returned if this is not the case. 
Clear any previously set centers via \code{clear_vor1_centers}.
}
