% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{set_vor2_active}
\alias{set_vor2_active}
\title{Set the active second-order Voronoi cells}
\usage{
VORobj$set_vor2_active(vor2_indices)
}
\arguments{
\item{vor2_indices}{a matrix whose rows contain the \code{(i,j)} indices (rows of \code{W}, 1-based) of the desired active second-order cells}
}
\value{
None, the field \code{vor2_active} is set internally.
}
\description{
The computations with a VOR object (computing Voronoi centers and ellipsoids) are only performed for "active" cells, 
which are identified by the prototype indices (row index in \code{W}) which generated the cell. 
By default the active second-order Voronoi cells are those which contain data mapped to them during a recall 
of the ANN, as indicated by CADJ (those second-order cells corresponding to \code{CADJ(i,j) > 0}). 
A different set of active cells can be set by calling this method
}
