% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{set_vor2_centers}
\alias{set_vor2_centers}
\title{Set the centers of second-order Voronoi cells}
\usage{
VORobj$set_vor2_centers
}
\arguments{
\item{C}{a matrix whose rows specify the active second-order Voronoi cell centers. Must have \code{nrow(C) = nrow(vor2_active)}.}
}
\value{
None, the field \code{vor2_centers} is set internally.
}
\description{
The MVIE and Dikin methods require a starting interior point for each active second-order Voronoi polytope, which 
can be set with this method. 
Note: the points passed to this method **must be interior** to each second-order cell; a check will 
be performed and an error returned if this is not the case. 
Clear any previously set centers via \code{clear_vor2_centers}.
}
