% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{SHGR}
\alias{SHGR}
\title{A SHGRWalk Synthetic Hyperspectral Image Cube}
\format{
For demonstration of Voronoi calculations in higher dimensions a 100-dimensional SHGR cube with 20 distinct classes (each with correlated noise) 
has been included in the \code{VorVQ} package. These data are stored in a list 
variable named \code{SHGR} with components: 
\describe{
\item{X}{data matrix whose 16,384 rows represent the 128 x 128 pixels in the image, 
and 100 columns represent the 100 spectral channels at which the synthetic reflectances were measured} 
\item{W}{matrix of 400 learned prototype vectors (in rows, learned from a 20x20 SOM)}
\item{CADJ}{The CADJ matrix resulting from a recall of the data in \code{X} using the prototypes in \code{W}. 
\code{CADJ} is ordered such that its \code{(i,j)} element contains the number of data vectors in \code{X} whose 
BMU1 = the prototype vector in the \code{i}-th row and BMU2 = the prototype in the \code{j}-th row of \code{W}}
}
}
\usage{
SHGR
}
\description{
The SHGRWalk (SHGR, for short) data suite is a collection of Synthetic Hyperspectral 128 x 128 pixel image cubes 
whose individual pixel "spectra" were sampled from a multi-component Gaussian Mixture Model whose component means 
were set via a (secondary) Gaussian Random Walk across the synthetic "spectral channels" (the data dimension, \code{d}). 
Individual component variances possesses a Toeplitz correlation structure with various noise levels. 
After sampling the pixels were labeled according to which mixture component (class) they were sampled from, 
and then organized into the 128 x 128 pixel image such that pixels from the same sampling component occupy 
contiguous blocks within the overall image.  
More information about the SHGRWalk data can be found here \code{need link}.
}
\keyword{datasets}
