% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_polytope.R
\name{vis_polytope}
\alias{vis_polytope}
\title{Visualize a (bounded) polytope}
\usage{
vis_polytope(
  A,
  b,
  center = NULL,
  add = F,
  fill.col = NULL,
  edge.lty = 1,
  edge.lwd = 1,
  edge.col = "navy",
  vertex.pch = 16,
  vertex.cex = 0.75,
  vertex.col = "navy",
  center.pch = 16,
  center.cex = 0.75,
  center.col = "navy",
  text = NULL,
  text.cex = 0.75,
  text.col = "navy"
)
}
\arguments{
\item{A}{the constraint matrix (must be 2-dimensional)}

\item{b}{the constraint RHS}

\item{center}{a flag denoting the type of center where the polytope label will be plotted as text. Can be one of
\enumerate{
\item NULL (default), no center defined
\item a vector (length = 2) giving the (x,y) coords of the center
\item "chebyshev", in which case the Chebyshev center will be calculated and used 
\item "analytic", in which case the analytic center will be calculated and used
}}

\item{add}{boolean, whether to generate a new plot or add to existing}

\item{fill.col}{color to fill polygon. Default = NULL means no fill used.}

\item{edge.lty}{edge linetype}

\item{edge.lwd}{edge linewidth}

\item{edge.col}{edge color}

\item{vertex.pch}{the vertex point style}

\item{vertex.cex}{the vertex point size, set = 0 to suppress vertex point plotting}

\item{vertex.col}{the vertex point color}

\item{center.pch}{the center point style}

\item{center.cex}{the center point size, set = 0 to suppress center point plotting if \code{center} != NULL}

\item{center.col}{the center point color}

\item{text}{a string (or numeric value, which will be cast as string) giving the label text which is plotted at (x,y) coords defined by \code{center}.
Default = NULL (nothing).}

\item{text.cex}{the text label size}

\item{text.col}{the text label color}
}
\value{
nothing, a plot is generated
}
\description{
The boundary region of a bounded polytope will be plotted as a set of intersecting line segments.
}
\details{
\code{center} must be given as something OTHER than NULL if either a point or label is requested to be plotted at the polytope center.
If only a text label at the center is desired, set center.cex = 0.
If only a point at the center is desired, set label = NULL
}
