% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_polytope.R
\name{vis_polytope_constraints}
\alias{vis_polytope_constraints}
\title{Visualize polytope constraints}
\usage{
vis_polytope_constraints(
  A,
  b,
  add = F,
  edge.lty = 1,
  edge.lwd = 1,
  edge.col = "magenta"
)
}
\arguments{
\item{A}{the constraint matrix (must be 2-dimensional)}

\item{b}{the constraint RHS}

\item{add}{boolean, whether to generate a new plot or add to existing}

\item{edge.lty}{edge linetype}

\item{edge.lwd}{edge linewidth}

\item{edge.col}{edge color}
}
\value{
nothing, a plot will be generated
}
\description{
The boundaries x : Ax = b will be plotted as lines
}
\details{
The plotting parameters *.lty, *.lwd and *.col are passed to R's plot() command. \cr
}
