% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_polytope.R
\name{vis_vor1_annotate}
\alias{vis_vor1_annotate}
\title{Annotate the first-order Voronoi cells}
\usage{
vis_vor1_annotate(
  VOR,
  add = F,
  text = "vorid",
  text.cex = 0.75,
  text.col = "black",
  text.font = 1
)
}
\arguments{
\item{VOR}{a Voronoi object, instantiated by \code{VOR$new()}}

\item{add}{boolean, whether to generate a new plot or add to existing}

\item{text}{the text label to be printed in each Voronoi cell. 
The default is a single string = 'vorid', which prints the ids in \code{$vor1_active} at the centers of each active first-order cell. 
Can also be given as a vector (length = length(\code{$vor1_active})) containing a label for each active first-order cell.}

\item{text.cex}{size of plotted text labels}

\item{text.col}{color for plotted text labels}

\item{text.font}{font weight for plotted text labels. Set = 2 for bold.}
}
\value{
nothing, a plot is generated
}
\description{
Annotate the first-order Voronoi cells
}
\details{
The text will be plotted at the \code{(x,y)} coordinates given by \code{$get_vor1_centers}.
}
