% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_polytope.R
\name{vis_vor1_Dikin}
\alias{vis_vor1_Dikin}
\title{Visualize the first-order Voronoi Dikin Ellipses}
\usage{
vis_vor1_Dikin(
  VOR,
  add = F,
  col = tableau20("green"),
  ell.lwd = 1.5,
  center.pch = 16,
  center.cex = 0.75
)
}
\arguments{
\item{VOR}{a Voronoi object, instantiated by \code{VOR$new()}}

\item{add}{boolean, whether to generate a new plot or add to existing}

\item{col}{color for plotting ellipse and its center}

\item{ell.lwd}{linewidth of ellipse}

\item{center.pch}{the center point style}

\item{center.cex}{the center point size, set = 0 to suppress center point plotting}
}
\value{
nothing, a plot is generated
}
\description{
Visualize the first-order Voronoi Dikin Ellipses
}
