% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_polytope.R
\name{vis_vor2_polytopes}
\alias{vis_vor2_polytopes}
\title{Visualize the second-order Voronoi cells generated by prototypes}
\usage{
vis_vor2_polytopes(
  VOR,
  add = F,
  edge.lty = 1,
  edge.lwd = 1,
  edge.col = tableau20("lightgray"),
  vertex.pch = 16,
  vertex.cex = 0.75,
  vertex.col = tableau20("lightgray")
)
}
\arguments{
\item{VOR}{a Voronoi object, instantiated by \code{VOR$new()}}

\item{add}{boolean, whether to generate a new plot or add to existing}

\item{edge.lty}{edge linetype}

\item{edge.lwd}{edge linewidth}

\item{edge.col}{edge color}

\item{vertex.pch}{the vertex point style}

\item{vertex.cex}{the vertex point size, set = 0 to suppress vertex point plotting}

\item{vertex.col}{the vertex point color}
}
\value{
nothing, a plot is generated
}
\description{
Visualize the second-order Voronoi cells generated by prototypes
}
