% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorobj_doc.R
\name{VOR}
\alias{VOR}
\title{The VOR object class}
\description{
The VOR object class
}
\section{Fields}{

\describe{
\item{\code{W}}{matrix of the prototypes (in rows) which generate the Voronoi tessellation}

\item{\code{nW}}{the number of prototypes in the tessellation}

\item{\code{d}}{the dimension of the prototypes \code{= ncols(W)}}

\item{\code{lb}}{a vector of dimension-wise lower bounds for the tessellated region, \code{length = d}}

\item{\code{ub}}{a vector of dimension-wise upper bounds for the tessellated region, \code{length = d}}

\item{\code{CADJ}}{the CADJ adjacency of the learned prototypes}

\item{\code{vor1_active}}{a vector of indices (rows of W) identifying the first-order Voronoi for which approximating ellipsoids will be computed. 
Defaults to the active first-order cells in the tessellation (those which contain at least one data observation after the ANN recall, as indicated by CADJ).}

\item{\code{vor2_active}}{a matrix whose rows contain the (i,j) indices identifying the second-order Voronoi cells for which approximating ellipsoids will be computed. 
Defaults to the active second-order cells in the tessellation (those which contain at least one data observation after the ANN recall, as indicated by CADJ).}

\item{\code{GADJ}}{the Gabriel graph adjacency matrix}

\item{\code{DADJ}}{the Delaunay graph adjacency matrix}

\item{\code{vor1_centers}}{matrix whose rows store the Chebyshev centers of the first-order Voronoi cells listed in \code{vor1_active}}

\item{\code{vor2_centers}}{matrix whose rows store the Chebyshev centers of the second-order Voronoi cells listed in \code{vor2_active}}

\item{\code{vor1_MVIE_c}}{matrix whose rows store the MVIE centers of the first-order Voronoi cells listed in \code{vor1_active}}

\item{\code{vor1_MVIE_E}}{cube whose slices store the MVIE ellipsoidal rotation matrices of the first-order Voronoi cells listed in \code{vor1_active}}

\item{\code{vor1_MVIE_logdet}}{vector whose elements store the log-determinant of MVIE ellipsoidal rotation matrices of the first-order Voronoi cells listed in \code{vor1_active}}

\item{\code{vor1_MVIE_volratio}}{vector whose elements store the proportional volume of the MVIEs of the first-order Voronoi cells listed in \code{vor1_active}}

\item{\code{vor1_MVIE_status}}{return flag from running the MVIE routine for each of the first-order Voronoi cells listed in \code{vor1_active}}

\item{\code{vor2_MVIE_c}}{matrix whose rows store the MVIE centers of the second-order Voronoi cells listed in \code{vor2_active}}

\item{\code{vor2_MVIE_E}}{cube whose slices store the MVIE ellipsoidal rotation matrices of the second-order Voronoi cells listed in \code{vor2_active}}

\item{\code{vor2_MVIE_logdet}}{vector whose elements store the log-determinant of MVIE ellipsoidal rotation matrices of the second-order Voronoi cells listed in \code{vor2_active}}

\item{\code{vor2_MVIE_volratio}}{vector whose elements store the proportional volume of the MVIEs of the second-order Voronoi cells listed in \code{vor2_active}}

\item{\code{vor2_MVIE_status}}{return flag from running the MVIE routine for each of the second-order Voronoi cells listed in \code{vor2_active}}

\item{\code{vor1_Dikin_E}}{cube whose slices store the Dikin ellipsoidal rotation matrices of the first-order Voronoi cells listed in \code{vor1_active}}

\item{\code{vor2_Dikin_E}}{cube whose slices store the Dikin ellipsoidal rotation matrices of the second-order Voronoi cells listed in \code{vor2_active}}

\item{\code{vis_par}}{List containing the \code{par()} parameters set during a call to the the \code{vis_vor_*} functions.}
}}

\section{Methods}{

Each class method has its own documentation, accessible via \code{?VorVQ::<method_name>}. 
For completeness, the list is repeated here in entirety.  Additional functionality 
for visualizing a two-dimensional VOR object is available through the \code{vis_*} functions. 
See their documentation for more information. 

\describe{
\item{\code{VOR$new}}{Instantiate a VOR object}
\item{\code{initialize_VOR}}{Initialize the VOR object with the prototypes and CADJ matrix from ANN learning}
\item{\code{set_bounds}}{Set the upper and lower bounds of the tessellated region}
\item{\code{set_vor1_active}}{Set the indices identifying the active first-order Voronoi cells}
\item{\code{set_vor2_active}}{Set the indices identifying the active second-order Voronoi cells}
\item{\code{get_params}}{Get the parameters used for various methods of the VOR object}
\item{\code{set_params}}{Set the parameters used for various methods of the VOR object}
\item{\code{calc_GADJ}}{Calculate the Gabriel adjacency matrix of the tessellation}
\item{\code{calc_DADJ}}{Calculate the Delaunay adjacency matrix of the tessellation}
\item{\code{set_DADJ}}{Set the Delaunay adjacency matrix of the tessellation, if obtained elsewhere}
\item{\code{get_params}}{Get the parameters used for various methods of the VOR object}
\item{\code{get_vor1_polytope}}{Get the polytope definition of a first-order Voronoi cell}
\item{\code{get_vor2_polytope}}{Get the polytope definition of a second-order Voronoi cell}
\item{\code{calc_vor1_centers}}{Compute the Chebhyshev centers of the active first-order Voronoi cells}
\item{\code{calc_vor2_centers}}{Compute the Chebhyshev centers of the active second-order Voronoi cells}
\item{\code{set_vor1_centers}}{Set the centers of the active first-order Voronoi cells}
\item{\code{set_vor2_centers}}{Set the centers of the active second-order Voronoi cells}
\item{\code{get_vor1_centers}}{Get the centers of the active first-order Voronoi cells}
\item{\code{get_vor2_centers}}{Get the centers of the active second-order Voronoi cells}
\item{\code{clear_vor1_centers}}{Clear the centers of the active first-order Voronoi cells}
\item{\code{clear_vor2_centers}}{Clear the centers of the active second-order Voronoi cells}
\item{\code{calc_vor1_MVIE}}{Compute the Maximum Volume Inscribed Ellipsoids (MVIE) of the active first-order Voronoi cells}
\item{\code{calc_vor2_MVIE}}{Compute the Maximum Volume Inscribed Ellipsoids (MVIE) of the active second-order Voronoi cells}
\item{\code{calc_vor1_Dikin}}{Compute the Dikin Ellipsoids of the active first-order Voronoi cells}
\item{\code{calc_vor2_Dikin}}{Compute the Dikin Ellipsoids of the active second-order Voronoi cells}
\item{\code{save}}{Save a VOR object to disk}
\item{\code{load}}{Load a previously saved VOR object from disk}
}
}

