% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vor1_polytope}
\alias{vor1_polytope}
\title{First Order Voronoi Cell Polytope
Returns the polytope definition Ax <= b of a single Voronoi cell}
\usage{
vor1_polytope(
  W,
  iidx,
  DADJ = NULL,
  lb = NULL,
  ub = NULL,
  rmv_redundancies = FALSE
)
}
\arguments{
\item{W}{matrix of prototypes in rows}

\item{iidx}{Index of Voronoi cell, 1-based.}

\item{DADJ}{(optional) - Delaunay adjacency. If used, the H-representation of the returned polytope will be tight, only including constraints from Delaunay neighbors.}

\item{lb}{(optional) - Global lower bound of data. If given, the H-representation of the returned polytope will include this as a lower bound.}

\item{ub}{(optional) - same as xlb, but for upper bounds. Both must be given in order for the global bounds to be appended to the polytope representation.}

\item{rmv_redundancies}{boolean, whether to check the system for redundancies, and remove any found}
}
\value{
a list with components A, b, cid
}
\description{
First Order Voronoi Cell Polytope
Returns the polytope definition Ax <= b of a single Voronoi cell
}
