% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vor2_polytope}
\alias{vor2_polytope}
\title{Second Order Voronoi Cell Polytope
Returns the polytope definition Ax <= b of a second order Voronoi cell}
\usage{
vor2_polytope(
  W,
  iidx,
  jidx,
  DADJ = NULL,
  lb = NULL,
  ub = NULL,
  rmv_redundancies = FALSE
)
}
\arguments{
\item{W}{matrix of prototypes in rows}

\item{iidx}{Index of 1st Voronoi cell}

\item{jidx}{Index of 2nd Voronoi cell}

\item{DADJ}{Delaunay adjacency. The H-representation of the returned polytope will be tight, only including constraints from Delaunay neighbors.}

\item{lb}{- Global lower bound of data. If given, the H-representation of the returned polytope will include this as a lower bound.}

\item{ub}{- same as xlb, but for upper bounds. Both must be given in order for the global bounds to be appended to the polytope representation.}

\item{rmv_redundancies}{boolean, whether to check the system for redundancies, and remove any found}
}
\value{
a list with components A, b, cid
}
\description{
Second Order Voronoi Cell Polytope
Returns the polytope definition Ax <= b of a second order Voronoi cell
}
